updated validations
