+++ backend/routes/api/users.js
@@ -3,55 +3,12 @@ const express = require('express')
 
 const { setTokenCookie } = require('../../utils/auth');
 const { User } = require('../../db/models');
-
 const { Op } = require("sequelize");
+const { validateSignup } = require('../../utils/inputValidators');
 
-const { check } = require('express-validator');
-const { handleValidationErrors } = require('../../utils/validation');
 
 
 const router = express.Router();
-
-const validateSignup = [
-    check('email')
-        .notEmpty()
-        .isEmail()
-        .withMessage('Email is invalid'),
-    check('username').isLength({min:4,max:30})
-            .withMessage('Username must be between 4 and 30 characters long'),
-    check('username')
-        .custom(username => {
-            if(typeof username !== 'string'){
-                throw new Error('Invalid Username')
-            }
-            return true;
-        })
-        .notEmpty()
-        .withMessage('Username is required'),
-    check('firstName').isLength({min:1,max:30})
-            .withMessage('First Name must be between 1 and 30 characters long'),
-    check('firstName')
-        .custom(firstName => {
-            if(typeof firstName !== 'string'){
-                throw new Error('Invalid First Name')
-            }
-            return true;
-        })
-        .notEmpty()
-        .withMessage('First Name is required'),
-    check('lastName').isLength({min:1,max:30})
-            .withMessage('Last Name must be between 1 and 30 characters long'),
-    check('lastName')
-        .custom(lastName => {
-            if(typeof lastName !== 'string'){
-                throw new Error('Invalid Last Name')
-            }
-            return true;
-        })
-        .notEmpty()
-        .withMessage('Last Name is required'),
-    handleValidationErrors
-];
 
 // Sign up
 router.post(
@@ -60,20 +17,6 @@ router.post(
     async (req, res) => {
 
         const { email, password, username, firstName, lastName } = req.body;
-
-        if(!username || !email){
-            res.status(400);
-            return res.json({
-                "message": "Validation error",
-                "statusCode": 400,
-                "errors": {
-                  "email": "Invalid email",
-                  "username": "Username is required",
-                  "firstName": "First Name is required",
-                  "lastName": "Last Name is required"
-                }
-            })
-        }
 
         let userWithUsernameOrEmail = await User.findOne({
             where: {
@@ -111,22 +54,7 @@ router.post(
         }else{
 
 
-            let user;
-            try{
-                user = await User.signup({ email, username, password, firstName, lastName });
-            }catch(err){
-                res.status(400);
-                return res.json({
-                    "message": "Validation error",
-                    "statusCode": 400,
-                    "errors": {
-                      "email": "Invalid email",
-                      "username": "Username is required",
-                      "firstName": "First Name is required",
-                      "lastName": "Last Name is required"
-                    }
-                });
-            }
+            let user = await User.signup({ email, username, password, firstName, lastName });
     
             const token = await setTokenCookie(res, user);
